+++ lib/representable/bindings/yaml_bindings.rb
@@ -34,15 +34,13 @@ module Representable
 
 
     class CollectionBinding < PropertyBinding
+      include Binding::Collection
+
       def serialize(value)
         Psych::Nodes::Sequence.new.tap do |seq|
           seq.style = Psych::Nodes::Sequence::FLOW if self[:style] == :flow
           value.each { |obj| seq.children << super(obj) }
         end
-      end
-
-      def deserialize(fragment)  # FIXME: redundant from Hash::Bindings
-        CollectionDeserializer.new(self).deserialize(fragment)
       end
     end
   end