+++ test/hash_bindings_test.rb
@@ -29,54 +29,6 @@ class HashBindingTest < MiniTest::Spec
       end
 
     end
-
-    describe "with plain text" do
-      before do
-        @property = Representable::Hash::PropertyBinding.new(Representable::Definition.new(:song), nil, nil)
-      end
-
-      it "extracts with #read" do
-        assert_equal "Thinning the Herd", @property.read("song" => "Thinning the Herd")
-      end
-
-      it "inserts with #write" do
-        doc = {}
-        assert_equal("Thinning the Herd", @property.write(doc,"Thinning the Herd"))
-        assert_equal({"song"=>"Thinning the Herd"}, doc)
-      end
-    end
-
-    describe "with an object" do
-      before do
-        @property = Representable::Hash::PropertyBinding.new(Representable::Definition.new(:song, :class => SongWithRepresenter), nil, nil)
-        @doc      = {}
-      end
-
-      it "extracts with #read" do
-        assert_equal SongWithRepresenter.new("Thinning the Herd"), @property.read("song" => {"name" => "Thinning the Herd"})
-      end
-
-      it "inserts with #write" do
-        assert_equal({"name"=>"Thinning the Herd"}, @property.write(@doc, SongWithRepresenter.new("Thinning the Herd")))
-        assert_equal({"song" => {"name"=>"Thinning the Herd"}}, @doc)
-      end
-    end
-
-    describe "with an object and :extend" do
-      before do
-        @property = Representable::Hash::PropertyBinding.new(Representable::Definition.new(:song, :class => Song, :extend => SongRepresenter), nil, nil)
-        @doc      = {}
-      end
-
-      it "extracts with #read" do
-        assert_equal Song.new("Thinning the Herd"), @property.read("song" => {"name" => "Thinning the Herd"})
-      end
-
-      it "inserts with #write" do
-        assert_equal({"name"=>"Thinning the Herd"}, @property.write(@doc, Song.new("Thinning the Herd")))
-        assert_equal({"song" => {"name"=>"Thinning the Herd"}}, @doc)
-      end
-    end
   end